Create tooltips in the item template instead
